@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>net7.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <Nullable>disable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
     <ApplicationInsightsResourceId>/subscriptions/beec1afc-268f-424a-aa60-34812bef200b/resourceGroups/HVOWebSiteV8/providers/microsoft.insights/components/HVOWebSiteV8</ApplicationInsightsResourceId>
@@ -15,7 +15,7 @@
   <ItemGroup>
 	  <PackageReference Include="Majorsoft.Blazor.Components.Loading" Version="1.5.0" />
 	  <PackageReference Include="Microsoft.ApplicationInsights.AspNetCore" Version="2.21.0" />
-    <PackageReference Include="Radzen.Blazor" Version="4.7.13" />
+    <PackageReference Include="Radzen.Blazor" Version="4.14.4" />
   </ItemGroup>
 
   <ItemGroup>
