@@ -1,31 +1,31 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
-  <PropertyGroup>
-    <TargetFramework>net7.0</TargetFramework>
-    <Nullable>disable</Nullable>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <ApplicationInsightsResourceId>/subscriptions/beec1afc-268f-424a-aa60-34812bef200b/resourceGroups/HVOv8/providers/microsoft.insights/components/HVOv8</ApplicationInsightsResourceId>
-    <UserSecretsId>448b2972-99a8-4816-b84f-393b365d821d</UserSecretsId>
-  </PropertyGroup>
+	<PropertyGroup>
+		<TargetFramework>net8.0</TargetFramework>
+		<Nullable>disable</Nullable>
+		<ImplicitUsings>enable</ImplicitUsings>
+		<ApplicationInsightsResourceId>/subscriptions/beec1afc-268f-424a-aa60-34812bef200b/resourceGroups/HVOv8/providers/microsoft.insights/components/HVOv8</ApplicationInsightsResourceId>
+		<UserSecretsId>448b2972-99a8-4816-b84f-393b365d821d</UserSecretsId>
+	</PropertyGroup>
 
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.OpenApi" Version="7.0.4" />
-	  <PackageReference Include="Microsoft.AspNetCore.Mvc.Versioning" Version="5.0.0" />
-	  <PackageReference Include="Microsoft.ApplicationInsights.AspNetCore" Version="2.21.0" />
-	  <PackageReference Include="Microsoft.AspNetCore.Mvc.Versioning.ApiExplorer" Version="5.0.0" />
-	  <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="7.0.4" />
-	  <PackageReference Include="Swashbuckle.AspNetCore" Version="6.5.0" />
-	  <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="7.0.4">
-		  <PrivateAssets>all</PrivateAssets>
-		  <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-	  </PackageReference>
+	<ItemGroup>
+		<PackageReference Include="Microsoft.AspNetCore.OpenApi" Version="8.0.0-preview.6.23329.11" />
+		<PackageReference Include="Microsoft.AspNetCore.Mvc.Versioning" Version="5.1.0" />
+		<PackageReference Include="Microsoft.ApplicationInsights.AspNetCore" Version="2.21.0" />
+		<PackageReference Include="Microsoft.AspNetCore.Mvc.Versioning.ApiExplorer" Version="5.1.0" />
+		<PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="8.0.0-preview.6.23329.4" />
+		<PackageReference Include="Swashbuckle.AspNetCore" Version="6.5.0" />
+		<PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="8.0.0-preview.6.23329.4">
+			<PrivateAssets>all</PrivateAssets>
+			<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+		</PackageReference>
 
-  </ItemGroup>
+	</ItemGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\HVO.DataContracts\HVO.DataContracts.csproj" />
-    <ProjectReference Include="..\HVO.DataModels\HVO.DataModels.csproj" />
-    <ProjectReference Include="..\HVO\HVO.csproj" />
-  </ItemGroup>
+	<ItemGroup>
+		<ProjectReference Include="..\HVO.DataContracts\HVO.DataContracts.csproj" />
+		<ProjectReference Include="..\HVO.DataModels\HVO.DataModels.csproj" />
+		<ProjectReference Include="..\HVO\HVO.csproj" />
+	</ItemGroup>
 
 </Project>
